+++ CVS History Viewer/Resources/Classes/DiffBlock.cs
@@ -7,6 +7,7 @@ namespace CVS_History_Viewer.Resources.Classes
     {
         public int iStartLine = 0;
         public int iEndLine = 0;
+        public string sBlockKind = "";
         public List<LineChange> cLines = new List<LineChange>();
         public List<TextBlock> cDiffLines = new List<TextBlock>();