+++ changelog.md
@@ -1,5 +1,21 @@
 # CVS History Viewer Changelog
 
+## v1.0.2 (Released 2019/05/21)
+### Improvements
+* (#7) Ignored files (any CVS related ones) will now only be checked once and then never again.
+* (#12) File checking logic is now faster.
+* (#14) Diff view now shows "Whitespace" (unchanged lines) in between diff blocks.
+* (#16) Line numbers in Diff view are now padded, so that they right-align with each other.
+* (#17) Diff view now uses the "Consolas" font, which is monospaced, meaning that each character takes up the same space.
+* (#18) App will now start-up a lot faster on big repositories, but commits will be limited to **220**. This limit can be overwritten by using the "limit" keyword (e.g. "limit:400").
+* (#19) Commit, author and date are now selectable.
+* (#20) Files that have been changed, but have no new commits, will now be checked again in the next refresh (even if no new change was made).
+### Bugfix
+* (#5) Progress bar is now filling up correctly.
+* (#6) "&"-Sign is now also being URL encoded and will no longer break the "Report Issue" URL.
+* (#13) Revision/Diff view is now reset, when no search results are found.
+* (#15) File, author and commit filter are now connected via "OR".
+
 ## v1.0.1 (Released 2019/04/29)
 ### Bugfix
 * (#1) Fixed a crash that were caused when analyzing diff with empty diff lines.